#9: extract logger to interface.
