+++ vault-upgrade-hook/pom.xml
@@ -174,5 +174,11 @@
             <version>1.7.7</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.assertj</groupId>
+            <artifactId>assertj-core</artifactId>
+            <version>2.8.0</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
