+++ vault-upgrade-hook/src/main/java/biz/netcentric/vlt/upgrade/UpgradeStatus.java
@@ -22,6 +22,7 @@ import org.apache.jackrabbit.vault.packaging.InstallContext;
 import org.apache.jackrabbit.vault.packaging.Version;
 
 import biz.netcentric.vlt.upgrade.util.PackageInstallLogger;
+import biz.netcentric.vlt.upgrade.util.PackageInstallLoggerImpl;
 
 /**
  * This class represents a previous upgrade execution and has methods to compare
@@ -29,7 +30,7 @@ import biz.netcentric.vlt.upgrade.util.PackageInstallLogger;
  */
 public class UpgradeStatus {
 
-    private static final PackageInstallLogger LOG = PackageInstallLogger.create(UpgradeStatus.class);
+    private static final PackageInstallLogger LOG = PackageInstallLoggerImpl.create(UpgradeStatus.class);
 
     public static final String PN_UPGRADE_TIME = "time";
     public static final String PN_VERSION = "version";
