+++ vault-upgrade-hook/src/main/java/biz/netcentric/vlt/upgrade/handler/script/ScriptHandler.java
@@ -15,10 +15,11 @@ import biz.netcentric.vlt.upgrade.UpgradeInfo;
 import biz.netcentric.vlt.upgrade.handler.SlingUtils;
 import biz.netcentric.vlt.upgrade.handler.UpgradeHandler;
 import biz.netcentric.vlt.upgrade.util.PackageInstallLogger;
+import biz.netcentric.vlt.upgrade.util.PackageInstallLoggerImpl;
 
 public class ScriptHandler implements UpgradeHandler {
 
-    private static final PackageInstallLogger LOG = PackageInstallLogger.create(ScriptHandler.class);
+    private static final PackageInstallLogger LOG = PackageInstallLoggerImpl.create(ScriptHandler.class);
 
     SlingUtils                                sling = new SlingUtils();
